update local setup

+++ connections/connection.go
@@ -2,5 +2,5 @@ package connections
 
 type Connection interface {
 	Connect()
-	ControlTemperature(temperature float64)
+	ControlTemperature(temperature uint32)
 }
